Separate views for control and index
