+++ webserver/views/__init__.py
@@ -1,0 +1,2 @@
+from webserver.views.ctrl import ctrl_view_blueprint
+from webserver.views.index import index_view_blueprint
